move .inhibitor related stuff
Move temporarily under the dbus module
